pyproject.toml
[tool.poetry]
name = "narrative-structures"
version = "0.1.0"
description = "Check narrative structure for consistency and proper creativeness"
authors = ["Iulian Pavlov <iulianpavlov@icloud.com>"]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-cors = "^5.0.0"
spacy = "^3.7.0"
ollama = "^0.3.3"
nltk = "^3.9.1"
python-telegram-bot = "^21.6"
python-dotenv = "^1.0.1"
flask-restx = "^1.3.0"
gigachat = "^0.1.35"
pydantic = ">=2.5.2"
python-docx = "^1.1.2"
flask-wtf = "^1.2.2"
gigachain-community = "^0.2.16"
pdfminer = "^20191125"
langchain-openai = ">=0.1.0"
langchain = ">=0.1.0"
langchain-core = ">=0.1.0"
langchain-community = ">=0.1.0"
langchain-anthropic = ">=0.1.0"
langchain-ollama = ">=0.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
